Causal Mediation Analysis

cmest is used to conduct causal mediation analysis via six statistical approaches including the regression-based approach by Valeri et al. (2013) and VanderWeele et al. (2014), the weighting-based approach by VanderWeele et al. (2014), the inverse odd-ratio weighting approach by Tchetgen Tchetgen (2013), the natural effect model by Vansteelandt et al. (2012), the marginal structural model by VanderWeele et al. (2017), and the g-formula approach by Lin et al. (2017).

cmest(
  data = NULL,
  model = NULL,
  full = TRUE,
  casecontrol = FALSE,
  yrare = NULL,
  yprevalence = NULL,
  estimation = "imputation",
  inference = "bootstrap",
  outcome = NULL,
  event = NULL,
  exposure = NULL,
  mediator = NULL,
  EMint = NULL,
  basec = NULL,
  postc = NULL,
  yreg = NULL,
  mreg = NULL,
  wmreg = NULL,
  ereg = NULL,
  postcreg = NULL,
  astar = 0,
  a = 1,
  mval = NULL,
  yref = NULL,
  basecval = NULL,
  nboot = 200,
  nRep = 5,
  multimp = FALSE,
  ...
)

# S3 method for cmest
print(x, ...)

# S3 method for cmest
summary(object, ...)

# S3 method for summary.cmest
print(x, digits = 4, ...)

# S3 method for cmsens.uc
print(x, ...)

# S3 method for cmsens.me
print(x, ...)

Arguments

data
dataset for causal mediation analysis.

model
causal mediation analysis approach. "rb", "wb", "iorw", "ne", "msm", "gformula" are implemented. See Details.

full
a logical value. If TRUE, output a full list of causal effects; if FALSE, output a reduced list of causal effects. Default is TRUE. See Details.

casecontrol
a logical value. TRUE indicates a case control study in which the first level of the outcome is treated as the control and the second level of the outcome is treated as the case. Default is FALSE.

yrare
a logical value (used when casecontrol is TRUE). TRUE indicates the case is rare.

yprevalence
the prevalence of the case (used when casecontrol is TRUE).

estimation
estimation method for causal effects. paramfunc and imputation are implemented (the first 4 letters are enough). Default is imputation. See Details.

inference
inference method for causal effects. delta and bootstrap are implemented (the first 4 letters are enough). Default is bootstrap. See Details.

outcome
variable name of the outcome.

event
variable name of the event (used when yreg is coxph, aft_exp, or aft_weibull).

exposure
variable name of the exposure.

mediator
a vector of variable name(s) of the mediator(s).

EMint
a logical value (used when yreg is character). If TRUE, the outcome regression formula includes the interaction(s) between the exposure and each of the mediator(s). Default is FALSE.

basec
a vector of variable name(s) of the baseline confounder(s)

postc
a vector of variable name(s) of mediator-outcome confounder(s) affected by the exposure

yreg
outcome regression. See Details.

mreg
a list of regression(s) specifying a regression model for each of the mediator(s) (used when model is rb, msm or gformula). See Details.

wmreg
a list of regression(s) specifying a regression model for calculating weights with respect to each of the mediator(s) (used when model is msm). See Details.

ereg
exposure regression for calculating weights with respect to the exposure (used when model is wb or msm with length(basec) != 0 or when model is iorw). See Details.

postcreg
a list of regression(s) specifying a regression model for each variable in postc (used when model is gformula). See Details.

astar
the first reference value for the exposure. Default is 0.

a
the second reference value for the exposure. Default is 1.

mval
a list specifying a reference value for each of the mediator(s) (used when model is rb, wb, ne, msm or gformula).

yref
reference value for the outcome (used when the outcome is categorical).

basecval
a list specifying a conditional value for each of the baseline confounder(s) for estimating conditional causal effects (used when estimation is paramfunc). If NULL, mean value(s) of the baseline confounder(s) are used.

nboot
the number of boots applied (used when inference is bootstrap). Default is 200.

nRep
number of replications or hypothetical values of the exposure to sample for each observation unit (used when model is ne). See neImpute.default for details. Default is 5.

multimp
a logical value (used when data contains missing values). If TRUE, conduct multiple imputation using the mice package. Default is FALSE.

...
Additional arguments passed to the mice function. See mice for details.

x
an object of class 'cmest'

object
an object of class 'cmest'

digits
minimal number of significant digits. See print.default.

Value

An object of class 'cmest' is returned:

call

the function call,

data

the dataset,

methods

a list of methods used which may include model, full, casecontrol, yprevalence, yrare, estimation, inference and nboot,

variables

a list of variables used which may include outcome, event, exposure, mediator, EMint, basec and postc,

reg.input

a list of regressions input,

multimp

a list of arguments used for multiple imputation,

ref

reference values used,

reg.output

a list of regressions output,

effect.pe

point estimates of causal effects,

effect.se

standard errors of causal effects,

effect.ci.low

the lower limits of confidence intervals of causal effects,

effect.ci.high

the higher limits of confidence intervals of causal effects,

effect.pe

point estimates of causal effects,

effect.pval

p-values of causal effects,

...
Details

Regressions

Each regression in yreg, mreg, wmreg, ereg and postcreg can be specified by a user-defined regression object or the character name of a regression.

The Character Name of A Regression

"linear": linear regression fitted by glm with family = gaussian()

"logistic": logistic regression fitted by glm with family = logit()

"loglinear": log linear regression fitted by glm with family = poisson() for a binary response

"poisson": poisson regression fitted by glm with family = poisson() for a count response

"quasipoisson": quasipoisson regression fitted by glm with family = quasipoisson()

"negbin": negative binomial regression fitted by glm.nb

"multinomial": multinomial regression fitted by multinom

"ordinal": ordered logistic regression fitted by polr

"coxph": cox proportional hazard model fitted by coxph

"aft_exp": accelerated failure time model fitted by survreg with dist = "exponential"

"aft_weibull": accelerated failure time model fitted by survreg with dist = "weibull"

coxph, aft_exp and aft_weibull are currently not implemented for mreg, wmreg, ereg and postcreg.

The User-defined Regression Object

A user-defined regression object can be fitted by lm, glm, glm.nb, gam, multinom, polr, coxph and survreg. Objects fitted by coxph and survreg are currently not supported for mreg, wmreg, ereg and postcreg.

The cmest function calculates weights for regressions when weighting is required. If a user-defined regression object is fitted with prior weights, the final weights for this regression are constructed by multiplying the prior weights with the weights calculated by the cmest function.

Causal Mediation Analysis Approaches

Let Y denote the outcome, A denote the exposure, M=(M_1,...,M_k)^T denote the set of mediator(s) following the temporal order, C denote the set of variables in basec, L=(L_1,...,L_s)^T denote the set of variables in postc following the temporal order.

rb: the regression-based approach by Valeri et al. (2013) and VanderWeele et al. (2014). yreg and mreg are required. If specified as a user defined regression object, yreg should regress Y on A, M and C and mreg[p] should regress M_p on A, M_1, ..., M_{p-1} and C for p=1,...,k.

wb: the weighting-based approach by VanderWeele et al. (2014). yreg is required. When basec is not empty, ereg is also required and A must be categorical. If specified as a user defined regression object, yreg should regress Y on A, M and C and ereg should regress A on C.

iorw: the inverse odd-ratio weighting approach by Tchetgen Tchetgen (2013). yreg and ereg are required and A must be categorical. If specified as a user defined regression object, yreg should regress Y on A and C and ereg should regress A on M and C.

ne: the natural effect model by Vansteelandt et al. (2012). yreg is required. If specified as a user defined regression object, yreg should regress Y on A, M and C. The variables in the formula of yreg should follow the order of A, M and C, i.e., the first variable needs to point to the exposure, the variable(s) right after the exposure need to point to the mediator(s), e.g., Y ~ A + M_1 + M_2 + A*M_1 + C.

msm: the marginal structural model by VanderWeele et al. (2017). yreg, mreg and wmreg are required and all mediators must be categorical. When basec is not empty, ereg is also required and A must be categorical. If specified as a user defined regression object, yreg should regress Y on A and M, mreg[p] should regress M_p on A, M_1, ..., and M_{p-1} for p=1,...,k, wmreg[p] should regress M_p on A, M_1, ..., M_{p-1}, C and L for p=1,...,k, and ereg should regress A on C.

gformula: the g-formula approach by Lin et al. (2017). yreg, mreg are required. postcreg is also required when postc is not empty. If specified as a user defined regression object, yreg should regress Y on A, M, C and L, mreg[p] should regress M_p on A, M_1, ..., M_{p-1}, C and L, for p=1,...,k, postcreg[p] should regress L_q on A, L_1, ..., L_{q-1} and C for q=1,...,s.

When postc is not empty, msm and gformula can be used, while rb, wb, iorw and ne can't be used.

Estimation Methods

paramfunc: closed-form parameter function estimation (only available when model = "rb" with length(mediator) = 1). The point estimate of each causal effect is obtained by its closed-form formula of regression coefficients. Effects conditional on pre-specified values of C, i.e., basecval, are estimated.

imputation: direct counterfactual imputation estimation. The point estimate of each causal effect is obtained by imputing counterfactuals directly.

To use closed-form parameter function estimation, yreg and mreg must be specified by the character name of the regression. yreg can be chosen from linear, logistic, loglinear, poisson, quasipoisson, negbin, coxph, aft_exp and aft_weibull. mreg can be chosen from linear, logistic and multinomial.

To use yreg = "logistic" in closed-form parameter function estimation, the binary outcome must be rare. For common binary outcomes, use yreg = "loglinear".

Inference Methods

delta: delta method (only available when estimation = "paramfunc"). The standard errors of causal effects are obtained by the delta method. The confidence intervals of causal effects are obtained by normal distribution approximation.

bootstrap: bootstrapping. The standard errors of causal effects are obtained by the standard deviations of bootstrapped results. The confidence intervals of causal effects are obtained by percentiles of bootstrapped results.

Estimated Causal Effects

When there are multiple mediators, the joint effect of them is studied.

For a continuous outcome, the causal effects on the difference scale are estimated. For a categorical, count or survival outcome, the causal effects on the ratio scale are estimated.

Continuous Outcome

When model = "rb", "wb", "ne", "msm" or "gformula" with length(postc) = 0 and full = TRUE, cde (controlled direct effect), pnde (pure natural direct effect), tnde (total natural direct effect), pnie (pure natural indirect effect), tnie (total natural indirect effect), te (total effect), intref (reference interaction), intmed (mediated interaction), cde(prop) (proportion cde), intref(prop) (proportion intref), intmed(prop) (proportion intmed), pnie(prop) (proportion pnie), pm (proportion mediated), int (proportion attributable to interaction) and pe (proportion eliminated) are estimated.

When model = "rb", "wb", "ne", "msm" or "gformula" with length(postc) = 0 and full = FALSE, cde, pnde, tnde, pnie, tnie and te are estimated.

When model = "msm" or "gformula" with length(postc) != 0 and full = TRUE, pnde, tnde, pnie, tnie, intref, intmed, intref(prop), intmed(prop), pie(prop), pm, int and pe are replaced by their randomized analogues rpnde, rtnde, rpnie, rtnie, rintref, rintmed, rintref(prop), rintmed(prop), rpie(prop), rpm, rint and rpe.

When model = "msm" or "gformula" with length(postc) != 0 and full = FALSE, cde, rpnde, rtnde, rpnie, rtnie and te are estimated.

When model = "iorw" and full = TRUE, te, pnde, tnie and pm are estimated. When model = "iorw" and full = FALSE, te, pnde and tnie are estimated.

Categorical, Count or Survival Outcome

When model = "rb", "wb", "ne", "msm" or "gformula" with length(postc) = 0 and full = TRUE, RRcde (cde risk rario or rate ratio), RRpnde (pnde risk rario or rate ratio), RRtnde (tnde risk rario or rate ratio), RRpnie (pnie risk rario or rate ratio), RRtnie (tnie risk rario or rate ratio), RRte (te risk rario or rate ratio), ERRcde (excess risk rario or rate ratio due to cde), ERRintref (excess risk rario or rate ratio due to intref), ERRintmed (excess risk rario or rate ratio due to intmed), ERRpie (excess risk rario or rate ratio due to pie), ERRcde(prop) (proportion ERRcde), ERRintref(prop) (proportion ERRintref), ERRintmed(prop) (proportion ERRintmed), ERRpie(prop) (proportion ERRpie), pm, int (proportion attributable to interaction) and pe (proportion eliminated) are estimated.

When model = "rb", "wb", "ne", "msm" or "gformula" with length(postc) = 0 and full = FALSE, RRcde, RRpnde, RRtnde, RRpnie, RRtnie and RRte are estimated.

When model = "msm" or "gformula" with length(postc) != 0 and full = TRUE, RRpnde, RRtnde, RRpnie, RRtnie, ERRintref, ERRintmed, ERRpie, ERRintref(prop), ERRintmed(prop), ERRpie(prop), pm, int and pe are replaced by their randomized analogues rRRpnde, rRRtnde, rRRpnie, rRRtnie, rERRintref, rERRintmed, rERRpie, rERRintref(prop), rERRintmed(prop), rERRpie(prop), rpm, rint and rpe.

When model = "msm" or "gformula" with length(postc) != 0 and full = FALSE, RRcde, rRRpnde, rRRtnde, rRRpnie, rRRtnie and RRte are estimated.

When model = "iorw" and full = TRUE, RRte, RRpnde, RRtnie and pm are estimated. When model = "iorw" and full = FALSE, RRte, RRpnde and RRtnie are estimated.

Methods (by generic)

print: Print the results of cmest nicely

summary: Summarize the results of cmest nicely

print: Print the summary of cmest nicely

print: Print the results of cmsens.uc nicely

print: Print the results of cmsens.me nicely

References

Valeri L, VanderWeele TJ (2013). Mediation analysis allowing for exposure-mediator interactions and causal interpretation: theoretical assumptions and implementation with SAS and SPSS macros. Psychological Methods. 18(2): 137 - 150.

VanderWeele TJ, Vansteelandt S (2014). Mediation analysis with multiple mediators. Epidemiologic Methods. 2(1): 95 - 115.

Tchetgen Tchetgen EJ (2013). Inverse odds ratio-weighted estimation for causal mediation analysis. Statistics in medicine. 32: 4567 - 4580.

Nguyen QC, Osypuk TL, Schmidt NM, Glymour MM, Tchetgen Tchetgen EJ. Practical guidance for conducting mediation analysis with multiple mediators using inverse odds ratio weighting (2015). American Journal of Epidemiology. 181(5): 349 - 356.

VanderWeele TJ, Tchetgen Tchetgen EJ (2017). Mediation analysis with time varying exposures and mediators. Journal of the Royal Statistical Society: Series B (Statistical Methodology). 79(3): 917 - 938.

Lin SH, Young J, Logan R, Tchetgen Tchetgen EJ, VanderWeele TJ (2017). Parametric mediational g-formula approach to mediation analysis with time-varying exposures, mediators, and confounders. Epidemiology. 28: 266 - 274.

Vansteelandt S, Bekaert M, Lange T. (2012). Imputation Strategies for the Estimation of Natural Direct and Indirect Effects. Epidemiologic Methods. 1(1): 131 - 158.

Steen J, Loeys T, Moerkerke B, Vansteelandt S (2017). Medflex: an R package for flexible mediation analysis using natural effect models. Journal of Statistical Software. 76(11).

VanderWeele TJ. A unification of mediation and interaction: a 4-way decomposition (2014). Epidemiology. 25(5): 749 - 61.

Imai K, Keele L, Tingley D. A general approach to causal mediation analysis (2010). Psychological Methods. 15(4): 309 - 334.

Schomaker M, Heumann C. Bootstrap inference when using multiple imputation (2018). Statistics in Medicine. 37(14): 2252 - 2266.

See also

Examples

if (FALSE) {
library(CMAverse)

# single-mediator case with rb
exp1 <- cmest(data = cma2020, model = "rb", outcome = "contY",
exposure = "A", mediator = "M2", basec = c("C1", "C2"),
EMint = TRUE, mreg = list("multinomial"), yreg = "linear",
astar = 0, a = 1, mval = list("M2_0"), estimation = "paramfunc",
inference = "delta")
summary(exp1)

# multiple-mediator case with rb
# 10 boots are used for illustration
exp2 <- cmest(data = cma2020, model = "rb", outcome = "contY",
exposure = "A", mediator = c("M1", "M2"), basec = c("C1", "C2"),
EMint = TRUE, mreg = list("logistic", "multinomial"),
yreg = "linear", astar = 0, a = 1, mval = list(0, "M2_0"),
estimation = "imputation", inference = "bootstrap", nboot = 10)

# multiple-mediator case with ne
# 10 boots are used for illustration
exp3 <- cmest(data = cma2020, model = "ne", outcome = "contY",
exposure = "A", mediator = c("M1", "M2"), basec = c("C1", "C2"),
yreg = glm(contY ~ A + M1 + M2 + A*M1 + A*M2 + C1 + C2, family = gaussian, data = cma2020),
astar = 0, a = 1, mval = list(0, "M2_0"), estimation = "imputation",
inference = "bootstrap", nboot = 10)

# case control study with msm
# 10 boots are used for illustration
exp4 <- cmest(data = cma2020, model = "msm", casecontrol = TRUE,
yrare = TRUE, outcome = "binY", exposure = "A",
mediator = c("M1", "M2"), EMint = TRUE, basec = c("C1", "C2"), yreg = "logistic",
ereg = "logistic", mreg = list(glm(M1 ~ A, family = binomial,
data = cma2020), nnet::multinom(M2 ~ A + M1, data = cma2020, trace = FALSE)),
wmreg = list(glm(M1 ~ A + C1 + C2, family = binomial, data = cma2020),
nnet::multinom(M2 ~ A + M1 + C1 + C2, data = cma2020, trace = FALSE)), astar = 0, a = 1,
mval = list(0, "M2_0"), estimation = "imputation",
inference = "bootstrap", nboot = 10)
}